+++ npm/react/plugins/load-webpack/index.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const returnInjectDevServerFunction = require('../utils/wrap-devserver')
+const returnInjectDevServerFunction = require('../utils/return-injectdevserver-function')
 const { startDevServer } = require('@cypress/webpack-dev-server')
 const tryLoadWebpackConfig = require('../utils/tryLoadWebpackConfig')
