+++ npm/react/plugins/react-scripts/index.js
@@ -1,4 +1,4 @@
-const returnInjectDevServerFunction = require('../utils/wrap-devserver')
+const returnInjectDevServerFunction = require('../utils/return-injectdevserver-function')
 const { startDevServer } = require('@cypress/webpack-dev-server')
 const findReactScriptsWebpackConfig = require('./findReactScriptsWebpackConfig')
