+++ npm/react/plugins/utils/return-injectdevserver-function.js
